# contrib/database/sql: move trace context info to w3c trace context tags