--- a/ddtrace/tracer/sqlcomment.go
+++ b/ddtrace/tracer/sqlcomment.go
@@ -86,6 +86,9 @@ func (c *SQLCommentCarrier) Inject(spanCtx ddtrace.SpanContext) error {
 			if v, ok := ctx.meta(ext.Version); ok {
 				version = v
 			}
+			if s := ctx.span; s != nil {
+				tags[sqlCommentDBService] = ctx.span.Service
+			}
 		}
 		if globalconfig.ServiceName() != "" {
 			tags[sqlCommentService] = globalconfig.ServiceName()