--- a/ddtrace/tracer/sqlcomment.go
+++ b/ddtrace/tracer/sqlcomment.go
@@ -86,7 +86,7 @@ func (c *SQLCommentCarrier) Inject(spanCtx ddtrace.SpanContext) error {
 			if v, ok := ctx.meta(ext.Version); ok {
 				version = v
 			}
-			if s := ctx.span; s != nil {
+			if ctx.span != nil {
 				tags[sqlCommentDBService] = ctx.span.Service
 			}
 		}