This makes an important update following the latest RFC proposals which adopts `traceparent` for trace context propagation instead of the ddtid, ddsid and ddsp tags. traceparent follows the [w3c trace context standard format](https://www.w3.org/TR/trace-context/#tracestate-header). 

A few related tweaks:
* I also renamed a few constants to match the more standard naming of "parent" instead of "client".
* I made the benchmark include the `Inject` in addition to the serialization of the comment. The reason being that `traceparent` includes some string formatting and I wanted to compare the performance before and after we switched to the formatted `traceparent`. You'll see that we get a hit from having a formatted value. It seems like an acceptable tradeoff. I did try an optimized version using string builder instead of `fmt.Sprintf` but it didn't seem to me like the difference was worth the tradeoff in readability (but curious what you'd think). You can see the optimized version in the commit history if curious.

Benchmark results (with a count of 10 to have a larger sample):
## Before `traceparent`
```
goos: darwin
goarch: amd64
pkg: gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer
cpu: Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz
BenchmarkSQLCommentInjection-8   	   10000	    121805 ns/op	  914701 B/op	      10 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    122675 ns/op	  914704 B/op	      10 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    117988 ns/op	  914703 B/op	      10 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    120263 ns/op	  914714 B/op	      10 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    120329 ns/op	  914712 B/op	      10 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    177295 ns/op	  914673 B/op	      10 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    172559 ns/op	  914685 B/op	      10 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    135637 ns/op	  914726 B/op	      10 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    119889 ns/op	  914719 B/op	      10 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    120622 ns/op	  914662 B/op	      10 allocs/op
```

## With `traceparent` using `fmt.Sprintf`:
```
goos: darwin
goarch: amd64
pkg: gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer
cpu: Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz
BenchmarkSQLCommentInjection-8   	   10000	    146027 ns/op	 1025596 B/op	      16 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    140450 ns/op	 1025594 B/op	      16 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    138458 ns/op	 1025591 B/op	      16 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    139865 ns/op	 1025600 B/op	      16 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    140585 ns/op	 1025597 B/op	      16 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    141854 ns/op	 1025597 B/op	      16 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    138637 ns/op	 1025600 B/op	      16 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    141157 ns/op	 1025598 B/op	      16 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    141207 ns/op	 1025594 B/op	      16 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    141695 ns/op	 1025596 B/op	      16 allocs/op
PASS
```

## With `traceparent` using `strings.Builder` (better results but enough to justify the tradeoff in readability?):
```
goos: darwin
goarch: amd64
pkg: gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer
cpu: Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz
BenchmarkSQLCommentInjection-8   	   10000	    138128 ns/op	 1025447 B/op	      15 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    135008 ns/op	 1025447 B/op	      15 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    128053 ns/op	 1025447 B/op	      15 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    134668 ns/op	 1025447 B/op	      15 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    133481 ns/op	 1025446 B/op	      15 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    132804 ns/op	 1025447 B/op	      15 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    131245 ns/op	 1025447 B/op	      15 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    131303 ns/op	 1025446 B/op	      15 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    133096 ns/op	 1025446 B/op	      15 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    134777 ns/op	 1025447 B/op	      15 allocs/op
PASS
```