@alexandre-normand 2022-09-02
> I think with unit tests covering us here, the benefits in performance from using the string builder seem worth it? A little loss in readability in an area of code that may end up being a bit of a hot path for certain customers / applications feels like the right choice. If there's any formatting or commenting we can do to offset then perhaps that will make it better?

That's fair. Let me bring that optimization back and we can review how well it reads and whether we can / should do anything about it.