@alexandre-normand 2022-09-02
> > I think with unit tests covering us here, the benefits in performance from using the string builder seem worth it? A little loss in readability in an area of code that may end up being a bit of a hot path for certain customers / applications feels like the right choice. If there's any formatting or commenting we can do to offset then perhaps that will make it better?
> 
> That's fair. Let me bring that optimization back and we can review how well it reads and whether we can / should do anything about it.

I made a `encodeTraceParent` function which helps a lot with readability. And while I was in there, I found another optimization to reduce allocations by growing the string Builder to a fixed length since we now exactly how many bytes traceparent is. Latest benchmark results:

```
goos: darwin
goarch: amd64
pkg: gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer
cpu: Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz
BenchmarkSQLCommentInjection-8   	   10000	    134344 ns/op	 1025391 B/op	      12 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    136646 ns/op	 1025391 B/op	      12 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    135072 ns/op	 1025391 B/op	      12 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    138164 ns/op	 1025391 B/op	      12 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    134253 ns/op	 1025391 B/op	      12 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    137485 ns/op	 1025391 B/op	      12 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    135041 ns/op	 1025391 B/op	      12 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    137130 ns/op	 1025391 B/op	      12 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    146462 ns/op	 1025391 B/op	      12 allocs/op
BenchmarkSQLCommentInjection-8   	   10000	    138535 ns/op	 1025391 B/op	      12 allocs/op
PASS
```